move utils to separate files. and add JSDocs
