+++ projects/schematics/src/shared/utils/variable-utils.ts
@@ -1,0 +1,44 @@
+import { findNodes } from '@schematics/angular/utility/ast-utils';
+import { parseTsFileContent } from './file-utils';
+import * as ts from 'typescript';
+
+interface ReplaceVariableDeclarationParams {
+  fileContent: string;
+  variableName: string;
+  newDeclaration: string;
+}
+
+/**
+ * Replaces a variable declaration in the given file content.
+ *
+ * Returns the updated file content.
+ */
+export function replaceVariableDeclaration({
+  fileContent,
+  variableName,
+  newDeclaration,
+}: ReplaceVariableDeclarationParams): string {
+  const sourceFile = parseTsFileContent(fileContent);
+
+  // Find all variable declarations
+  const nodes = findNodes(sourceFile, ts.SyntaxKind.VariableDeclaration);
+
+  // Find the specific variable declaration we want to replace
+  const targetNode = nodes.find((node) => {
+    if (!ts.isVariableDeclaration(node)) return false;
+    const name = node.name;
+    return ts.isIdentifier(name) && name.text === variableName;
+  });
+
+  if (!targetNode) {
+    return fileContent;
+  }
+
+  // Get the parent VariableStatement to include the 'const' keyword
+  const statement = targetNode.parent.parent;
+  const start = statement.getStart();
+  const end = statement.getEnd();
+
+  // Replace the entire statement
+  return fileContent.slice(0, start) + newDeclaration + fileContent.slice(end);
+}
